Move string.print -> io.print
