+++ koto/benches/n_body.koto
@@ -8,7 +8,7 @@ https://benchmarksgame-team.pages.debian.net/benchmarksgame/program/nbody-lua-4.
 
 # TODO Using num4 for pos/vel reduces precision compared to the original
 
-import number.pi, test.assert_near
+import io.print, number.pi, test.assert_near
 
 solar_mass = 4 * pi * pi
 days_per_year = 365.24
@@ -124,8 +124,8 @@ export main = ||
   quiet = (koto.args.get 1) == "quiet"
 
   if not quiet
-    "{:.9}".print initial_energy
-    "{:.9}".print end_energy
+    print "{:.9}", initial_energy
+    print "{:.9}", end_energy
 
 export @tests =
   @test nbody_100: ||
